+++ src/runtime/mod.rs
@@ -880,6 +880,17 @@ impl<'rant> VM<'rant> {
           self.push_getter_intents(path, false, false, fallback.as_ref().map(Rc::clone));
           return Ok(true)
         },
+        Rst::VarDepth(vname, access_kind, fallback) => {
+          match (self.get_var_depth(vname, *access_kind), fallback) {
+            (Ok(depth), _) => self.cur_frame_mut().write_value(RantValue::Int(depth as i64)),
+            (Err(_), Some(fallback)) => {
+              self.cur_frame_mut().push_intent_front(Intent::PrintLast);
+              self.push_frame(Rc::clone(fallback), true)?;
+              return Ok(true)
+            },
+            (Err(err), None) => return Err(err),
+          }
+        },
         Rst::VarSet(path, val_expr) => {
           // Get list of dynamic expressions in path
           let exprs = path.dynamic_exprs();
@@ -1505,6 +1516,11 @@ impl<'rant> VM<'rant> {
   #[inline(always)]
   pub fn get_var_value(&self, varname: &str, access: AccessPathKind, prefer_function: bool) -> RuntimeResult<RantValue> {
     self.call_stack.get_var_value(self.engine, varname, access, prefer_function)
+  }
+
+  #[inline(always)]
+  pub fn get_var_depth(&self, varname: &str, access: AccessPathKind) -> RuntimeResult<usize> {
+    self.call_stack.get_var_depth(self.engine, varname, access)
   }
 
   /// Defines a new variable in the current scope.