+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New
 * Added new number format system: `alpha`
+* Added "depth operator" to determine originating scope of a variable
 
 ### Changes
 * `[seg]` now operates internally on grapheme boundaries instead of `char` boundaries